backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Church Calendar Backend"
authors = [
    {name = "Lazaro Altedill",email = "lazaroah02@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "markdown (>=3.8,<4.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
